Former-commit-id: 71683dbb5cd689c590dc8ea89d8be203c4d91431
Former-commit-id: 3d30ebb3424c0ad53472a4068b1724651501c451

+++ Ignite-Core/include/Ignite/Renderer/Camera.h
@@ -18,7 +18,7 @@ namespace Ignite
 
 		void Translate(CameraDirection direction, float speed);
 
-		void MousePosition(double xpos, double ypos);
+		void MousePosition(float xpos, float ypos);
 
 		glm::mat4 GetViewMatrix() const;
